updated nav styles
