+++ src/components/nav.tsx
@@ -4,13 +4,13 @@ import { HamburgerMenuIcon } from '@radix-ui/react-icons';
 
 export default function Nav() {
 return (
-    <NavigationMenu.Root className='NavigationMenuRoot mobile-nav' delayDuration={999999}>
-        <NavigationMenu.List className='NavigationMenuList nav-links'>
-            <NavigationMenu.Item>
+    <NavigationMenu.Root className='NavigationMenuRoot w-full absolute z-10 bg-light' delayDuration={99999999}>
+        <NavigationMenu.List className='NavigationMenuList border-b border-dark'>
+            <NavigationMenu.Item className='h-fit flex flex-col'>
                 <NavigationMenu.Trigger className='NavigationMenuTrigger'>
-                    <HamburgerMenuIcon className='nav-icon' width={40} height={40}/>
+                    <HamburgerMenuIcon className='m-2' width={40} height={40}/>
                 </NavigationMenu.Trigger>
-                <NavigationMenu.Content className='NavigationMenuContent'>
+                <NavigationMenu.Content className='NavigationMenuContent flex flex-col w-full'>
                     <NavigationMenu.Link asChild>
                         <Link href='/' className='nav nav-link nav-home'>
                             <p className='nav-p'>Home</p>
@@ -29,7 +29,7 @@ return (
                     <NavigationMenu.Link asChild>
                         <Link
                         href='https://github.com/heyligerjon'
-                        className='nav nav-link nav-github'
+                        className='nav-link nav-github'
                         target='_blank'
                         rel='noopener noreferrer'
                         >
@@ -37,13 +37,18 @@ return (
                         </Link>
                     </NavigationMenu.Link>
                     <NavigationMenu.Link asChild>
-                        <Link href='#contact' className='nav nav-link nav-contact'>
+                        <Link href='/#contact' className='nav nav-link nav-contact'>
                             <p className='nav-p'>Contact</p>
                         </Link>
                     </NavigationMenu.Link>
                     <NavigationMenu.Link asChild>
                         <Link href='experience' className='nav nav-link nav-exp'>
                             <p className='nav-p'>View My Resume</p>
+                        </Link>
+                    </NavigationMenu.Link>
+                    <NavigationMenu.Link asChild>
+                        <Link href='' className=''>
+                            <p className=''> </p>
                         </Link>
                     </NavigationMenu.Link>
                 </NavigationMenu.Content>
