+++ src/styles/globals.css
@@ -39,8 +39,18 @@ input, textarea {
 }
 
 /* THEMING, ANIMATIONS, & COLORS */
+.bg-light {
+  background-color: #f8f5eb;
+  color: #021e36;
+}
+
+.bg-dark {
+  background-color: #021e36;
+  color: #f8f5eb;
+}
+
 .border-dark {
-  border: 2px solid #021e36;
+  border-color: #021e36;
 }
 
 .gradient {
@@ -108,22 +118,17 @@ code {
 }
 
 /* COMPONENTS */
-.mobile-nav {
-  width: 100%;
-  transition: linear ease-in 0.5s;
-}
-.mobile-nav div {
-  width: 100%;
-}
-.mobile-nav .nav-icon {
-  margin: 0.5rem 0;
-}
-.mobile-nav .nav-p {
-  height: fit-content;
+.nav-link {
   display: flex;
   justify-content: center;
-  margin-bottom: 0.5rem;
+  padding: 0.25rem;
   width: 100%;
+  height: 100%;
+}
+
+.nav-link:hover, .nav-link:focus, .nav-link:active {
+  background-color: #021e36;
+  color: #f8f5eb;
 }
 
 /* HOMEPAGE STYLES */
@@ -369,13 +374,6 @@ footer {
   .container {
     max-width: 100vw;
   }
-  .mobile-nav {
-    display: flex;
-  }
-  .mobile-nav .nav-links li {
-    background-color: #f8f5eb;
-    border-bottom: 2px solid #021e36;
-  }
   .exp-tabs {
     display: flex;
     justify-content: center;
@@ -432,16 +430,6 @@ footer {
   }
   .container {
     max-width: 100vw;
-  }
-  .mobile-nav {
-    display: flex;
-  }
-  .mobile-nav .nav-links {
-    background-color: #f8f5eb;
-    border-bottom: 2px solid #021e36;
-  }
-  .home-nav {
-    display: none;
   }
   .exp-tabs {
     display: flex;
@@ -503,24 +491,6 @@ footer {
   .container {
     max-width: 80vw;
   }
-  .mobile-nav {
-    display: none;
-  }
-  .home-nav {
-    display: flex;
-    position: relative;
-    width: 33%;
-  }
-  .home-nav .nav-item {
-    border-radius: 30px 0 0 30px;
-    width: 100%;
-  }
-  .home-nav .nav-item a {
-    width: fit-content;
-  }
-  .home-nav .nav-item .nav-p {
-    padding-left: 2rem;
-  }
   .exp-tabs {
     display: flex;
     justify-content: center;
@@ -577,23 +547,6 @@ footer {
   input, textarea {
     width: 600px;
   }
-  .mobile-nav {
-    display: none;
-  }
-  .home-nav {
-    display: flex;
-    position: relative;
-  }
-  .home-nav .nav-item {
-    border-radius: 30px 0 0 30px;
-    width: 100%;
-  }
-  .home-nav .nav-item a {
-    width: fit-content;
-  }
-  .home-nav .nav-item .nav-p {
-    padding-left: 2rem;
-  }
   .exp-tabs {
     display: flex;
     justify-content: center;
@@ -644,26 +597,6 @@ footer {
   }
   input, textarea {
     width: 600px;
-  }
-  .mobile-nav {
-    display: none;
-  }
-  .home-nav {
-    display: flex;
-    position: relative;
-    width: 35%;
-  }
-  .home-nav .nav-item {
-    margin-left: 0rem;
-    border-right: 2px solid #021e36;
-    border-radius: 30px;
-    width: 75%;
-  }
-  .home-nav .nav-item a {
-    width: 100%;
-  }
-  .home-nav .nav-item .nav-p {
-    padding-left: 5em;
   }
   .exp-tabs {
     display: flex;
